requirements.txt
playwright>=1.40.0
anthropic>=0.18.0
openai>=1.12.0
tiktoken>=0.6.0
python-dotenv>=1.0.0
